# dist/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": { "title": "COD Platform API", "version": "1.0.0" },
    "servers": [{ "url": "https://cod-platform.onrender.com" }],
    "components": {
        "securitySchemes": {
            "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
        }
    },
    "security": [],
...
    "paths": {
...
        "/api/auth/refresh": {
            "post": {
                "tags": ["Auth"],
                "summary": "Refresh access token",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": { "refreshToken": { "type": "string" } }
                            }
                        }
                    }
                },
                "responses": { "200": { "description": "OK" } }
            }
        },
...
    }
}